requirements.txt
python-dotenv==1.0.1
schedule==1.2.1
requests==2.31.0
